python/weather/requirements.txt
asgiref==3.7.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
distlib==0.3.7
Django==4.2.5
extension==1.0.0
filelock==3.12.4
Flask==2.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.4
MarkupSafe==2.1.3
pipenv==2023.9.8
platformdirs==3.10.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.24.5
Werkzeug==2.3.7
